+++ src/classes/ValidationError.ts
@@ -1,0 +1,6 @@
+
+/**
+ * An error which is thrown if a validation fails.
+ */
+export class ValidationError extends Error {
+}